+++ R/plot.dispRity_fun.R
@@ -1073,6 +1073,7 @@ plot.model.sim <- function(data, add, density, quantiles, cent.tend, ...) {
     plot.continuous(plot_params, data_params, add = add, density = density)
 }
 
+## Plotting test metrics
 plot.test.metric <- function(data, specific.args, ...) {
 
     ## Adding slopes
@@ -1389,4 +1390,88 @@ plot.test.metric <- function(data, specific.args, ...) {
     if(n_plots > 1 && !is.null(op_tmp)) {
         par(op_tmp)
     }
+}
+
+## Plot axes
+plot.axes <- function(data, ...) {
+
+    ## Magic value for below
+    transparency <- 0.5
+
+    ## Get the number of groups
+    n_groups <- length(data$dim.list)
+
+    ## Get the columns to plot
+    if(is.null(data$call$colnames)) {
+        names.arg <- paste0("dim.", 1:length(data$scaled.var[[1]]))
+    } else {
+        names.arg <- data$call$colnames
+    }
+
+    ## Plotting windows
+    op_tmp <- par(mfrow = c(ceiling(sqrt(n_groups)), round(sqrt(n_groups))))
+
+    ## Getting the plotting parameters
+    plot_params <- list(...)
+    ## Colour
+    if(is.null(plot_params$col)) {
+        plot_params$col <- rep("grey", n_groups)
+    } else {
+        if(length(plot_params$col) < n_groups) {
+            plot_params$col <- rep(plot_params$col, n_groups)
+        }
+    }
+    ## Main
+    if(is.null(plot_params$main)) {
+        plot_params$main <- names(data$dim.list)
+    } else {
+        if(length(plot_params$main) < n_groups) {
+            plot_params$main <- rep(plot_params$main, n_groups)
+        }
+    }
+    ## Ylab
+    if(is.null(plot_params$ylab)) {
+        plot_params$ylab <- "Scaled variance"
+    }
+
+    ## Get the max number of dimensions
+    max_dim <- length(data$scaled.var[[1]])
+    ## Xlim (selecting the maximum number of bars)
+    if(is.null(plot_params$xlim)) {
+        if(max(data$dimensions) != max_dim) {
+            ## Only display the selected dimensions + 15%
+            display <- ceiling(max(data$dimensions)*1.1)
+            if(display <= max_dim) {
+                max_dim <- display
+            }
+        }
+    } else {
+        display <- max(plot_params$xlim)
+        if(display <= max_dim) {
+            max_dim <- display
+        }
+        plot_params$xlim <- NULL
+    }
+
+    ## Plotting the bars
+    for(one_group in 1:n_groups) {
+        ## Cum bars
+        bar_params <- plot_params
+        bar_params$col <- grDevices::adjustcolor(bar_params$col[one_group], alpha.f = transparency)
+        bar_params$main <- bar_params$main[one_group]
+        bar_params$height <- data$cumsum.var[[one_group]][1:max_dim]
+        bar_params$names.arg <- names.arg[1:max_dim]
+        bars_coords <- do.call(barplot, bar_params)
+        ## Scaled bars
+        bar_params <- plot_params
+        bar_params$col <- bar_params$col[one_group]
+        bar_params$height <- data$scaled.var[[one_group]][1:max_dim]
+        bar_params$add <- TRUE
+        do.call(barplot, bar_params)
+        ## Add the threshold line
+        abline(h = data$call$threshold, lty = 2)
+        abline(v = bars_coords[max(data$dim.list[[one_group]])], lty = 2)
+    }
+    par(op_tmp)
+    return(invisible())
 }