+++ NEWS.md
@@ -23,12 +23,17 @@ dispRity 0.2.0 (2016-04-01)
 <!--   * `dtt.dispRity` now works with time-slicing. -->
 
 
-dispRity v1.6.3 (2021-07-01)
+dispRity v1.6.4 (2021-08-02)
 =========================
 
 ### NEW FEATURES
 
+ * *New* data function: `select.axes` for selecting and analysing the number of axes required to contain an arbitrary amount of variance.
  * *New* utility function: `randtest.dist` for measuring the distance between the observed statistic and a specific quantile of the simulated statistic (thanks to [Frane Babarovic](https://twitter.com/FBabarovic) for the inspiration).
+
+### MINOR IMPROVEMENTS
+
+ * formalised the grouping logic for `custom.subsets` and `select.axes`. This can create some minor user level changes namely: warning messages for empty subsets now correctly mentions "subsets" (rather than subsamples); groups with incorrect elements are now always flagged as errors (rather than just ignored). The changes at the developer level is that the logic is now made smoother and exported in `custom.subsets_fun.R`.
 
 ### BUG FIXES